Extract aggregation code to general utils file
